refactor: 💡 [api] remove repository layer
